[AppleHealth] 实现体重BMI管理功能 - 新增 weight_bmi.go 分析模块   - 体重趋势追踪   - BMI计算和分类(偏瘦/正常/超重/肥胖)   - 统计数据(最小/最大/平均体重,体重变化)   - 30天体重预测(线性回归)   - 个性化健康建议 - 新增 API 端点 GET /api/v1/health/analysis/weight-bmi - 修复 screentime 模块类型转换警告

+++ internal/screentime/advanced_analysis.go
@@ -880,7 +880,7 @@ func (h *Handler) generateWeekdayWeekendInsights(comparison WeekdayWeekendCompar
 	weekdayWork := comparison.Weekday.CategoryDistribution["Productivity"] + comparison.Weekday.CategoryDistribution["Business"]
 	weekendWork := comparison.Weekend.CategoryDistribution["Productivity"] + comparison.Weekend.CategoryDistribution["Business"]
 
-	if weekendWork > weekdayWork*0.5 {
+	if weekendWork > int64(float64(weekdayWork)*0.5) {
 		insights = append(insights, "周末仍有较多工作相关应用使用，建议注意工作生活平衡")
 	}
